configured the api by including StoreContext, updating CatalogController, fixed CatalogController in directory and renamed, added Id property to Ratings, excluded database files
