+++ src/api/searching.troll.Api/Program.cs
@@ -1,7 +1,16 @@
 using Microsoft.OpenApi.Models;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Sqlite;
+using searching.troll.Data;
+using Microsoft.Extensions.Options;
 
 var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddControllers();
+builder.Services.AddDbContext<StoreContext>(options =>
+{
+    options.UseSqlite("Data Source=../store.db",
+    m => m.MigrationsAssembly("searching.troll.Api"));
+});
 
 builder.Services.AddSwaggerGen(c =>
 {
