+++ src/api/searching.troll.Api/searching.troll.Api.csproj
