+++ src/api/searching.troll.Domain/Catalog/Rating.cs
@@ -2,6 +2,7 @@ namespace searching.troll.Domain.Catalog;
 
 public class Rating
 {
+    public int Id { get; set; }
 
     public int Stars { get; set; }
     public string UserName { get; set; }
